# openapi-check.json
{"openapi":"3.0.3","info":{"title":"Travel Plan App API","description":"REST API for creating, managing, and publishing travel itineraries with HTML rendering","version":"1.0.0","contact":{"name":"Travel Plan App"}},"servers":[{"url":"http://localhost:3000","description":"Development server"}],"tags":[{"name":"trips","description":"Trip management endpoints"},{"name":"documentation","description":"API documentation"}],"paths":{"/api/trips":{"get":{"tags":["trips"],"summary":"List all trips","description":"Returns a list of all trips with optional search","parameters":[{"name":"q","in":"query","description":"Search query for trip titles and stop names","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"type":"object","properties":{"trips":{"type":"array","items":{"$ref":"#/components/schemas/TripSummary"}},"query":{"type":"string"}}}}}},"500":{"$ref":"#/components/responses/InternalServerError"}}},"post":{"tags":["trips"],"summary":"Create new trip","description":"Creates a new trip with generated ID and timestamps","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTripInput"}}}},"responses":{"201":{"description":"Trip created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Trip"}}}},"400":{"$ref":"#/components/responses/ValidationError"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},"/api/trips/{id}":{"get":{"tags":["trips"],"summary":"Get trip by ID","description":"Returns full trip details","parameters":[{"$ref":"#/components/parameters/TripId"}],"responses":{"200":{"description":"Successful response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Trip"}}}},"404":{"$ref":"#/components/responses/NotFound"},"500":{"$ref":"#/components/responses/InternalServerError"}}},"patch":{"tags":["trips"],"summary":"Update trip","description":"Partially updates a trip (merge semantics)","parameters":[{"$ref":"#/components/parameters/TripId"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTripInput"}}}},"responses":{"200":{"description":"Trip updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Trip"}}}},"400":{"$ref":"#/components/responses/ValidationError"},"404":{"$ref":"#/components/responses/NotFound"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},"/api/trips/{id}/render":{"get":{"tags":["trips"],"summary":"Render trip as HTML","description":"Returns server-side rendered HTML page for browser viewing","parameters":[{"$ref":"#/components/parameters/TripId"}],"responses":{"200":{"description":"HTML rendered successfully","content":{"text/html":{"schema":{"type":"string"}}}},"404":{"description":"Trip not found","content":{"text/html":{"schema":{"type":"string","example":"<h1>Trip not found</h1>"}}}},"500":{"description":"Rendering error","content":{"text/html":{"schema":{"type":"string","example":"<h1>Error rendering trip</h1>"}}}}}}},"/api/trips/{id}/publish":{"post":{"tags":["trips"],"summary":"Publish trip as static HTML","description":"Generates and saves static HTML file in public directory","parameters":[{"$ref":"#/components/parameters/TripId"}],"responses":{"200":{"description":"Trip published successfully","content":{"application/json":{"schema":{"type":"object","properties":{"url":{"type":"string","example":"/published/550e8400-e29b-41d4-a716-446655440000.html"},"path":{"type":"string"}}}}}},"404":{"$ref":"#/components/responses/NotFound"},"500":{"$ref":"#/components/responses/InternalServerError"}}}},"/api/openapi.json":{"get":{"tags":["documentation"],"summary":"Get OpenAPI specification","description":"Returns the OpenAPI 3.0 specification for this API","responses":{"200":{"description":"OpenAPI spec","content":{"application/json":{"schema":{"type":"object"}}}}}}}},"components":{"parameters":{"TripId":{"name":"id","in":"path","description":"Trip UUID","required":true,"schema":{"type":"string","format":"uuid"}}},"schemas":{"Trip":{"type":"object","required":["id","title","days","createdAt","updatedAt"],"properties":{"id":{"type":"string","format":"uuid"},"title":{"type":"string","minLength":1},"startDate":{"type":"string","format":"date","pattern":"^\\d{4}-\\d{2}-\\d{2}$"},"endDate":{"type":"string","format":"date","pattern":"^\\d{4}-\\d{2}-\\d{2}$"},"travelers":{"$ref":"#/components/schemas/Travelers"},"preferences":{"$ref":"#/components/schemas/Preferences"},"days":{"type":"array","items":{"$ref":"#/components/schemas/Day"}},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"TripSummary":{"type":"object","required":["id","title","updatedAt"],"properties":{"id":{"type":"string","format":"uuid"},"title":{"type":"string"},"updatedAt":{"type":"string","format":"date-time"}}},"CreateTripInput":{"type":"object","required":["title"],"properties":{"title":{"type":"string","minLength":1},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"travelers":{"$ref":"#/components/schemas/Travelers"},"preferences":{"$ref":"#/components/schemas/Preferences"},"days":{"type":"array","items":{"$ref":"#/components/schemas/Day"}}}},"UpdateTripInput":{"type":"object","properties":{"title":{"type":"string","minLength":1},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"travelers":{"$ref":"#/components/schemas/Travelers"},"preferences":{"$ref":"#/components/schemas/Preferences"},"days":{"type":"array","items":{"$ref":"#/components/schemas/Day"}}}},"Day":{"type":"object","required":["dayNumber","start","end","stops"],"properties":{"dayNumber":{"type":"integer","minimum":1},"date":{"type":"string","format":"date"},"start":{"$ref":"#/components/schemas/Place"},"end":{"$ref":"#/components/schemas/Place"},"driveEstimate":{"$ref":"#/components/schemas/DriveEstimate"},"stops":{"type":"array","items":{"$ref":"#/components/schemas/Stop"},"default":[]},"notes":{"type":"string"},"googleMapsUrl":{"type":"string","format":"uri","description":"URL to Google Maps route for this day"}}},"Place":{"type":"object","required":["name"],"properties":{"name":{"type":"string","minLength":1},"lat":{"type":"number"},"lng":{"type":"number"},"address":{"type":"string"},"imageUrl":{"type":"string","format":"uri","description":"URL to image of this place"}}},"Stop":{"type":"object","required":["place"],"properties":{"timeBlock":{"type":"string","enum":["morning","midday","afternoon","evening"]},"place":{"$ref":"#/components/schemas/Place"},"category":{"type":"string","enum":["scenic","food","culture","hike","lodging","misc"]},"durationMinutes":{"type":"integer","minimum":1},"costLevel":{"type":"integer","enum":[1,2,3,4]},"booking":{"$ref":"#/components/schemas/Booking"},"notes":{"type":"string"},"imageUrl":{"type":"string","format":"uri","description":"URL to image for this specific stop"},"externalLinks":{"type":"array","items":{"$ref":"#/components/schemas/ExternalLink"},"description":"Related websites, guides, or resources"}}},"ExternalLink":{"type":"object","required":["title","url"],"properties":{"title":{"type":"string","minLength":1,"description":"Display text for the link"},"url":{"type":"string","format":"uri","description":"External URL"}}},"Booking":{"type":"object","properties":{"url":{"type":"string","format":"uri"},"confirmation":{"type":"string"}}},"DriveEstimate":{"type":"object","properties":{"hours":{"type":"number","minimum":0},"km":{"type":"number","minimum":0}}},"Travelers":{"type":"object","properties":{"adults":{"type":"integer","minimum":0},"kids":{"type":"integer","minimum":0}}},"Preferences":{"type":"object","properties":{"pace":{"type":"string","enum":["relaxed","balanced","packed"]},"dailyDriveHoursMax":{"type":"number","minimum":0},"interests":{"type":"array","items":{"type":"string"}},"avoid":{"type":"array","items":{"type":"string"}},"budget":{"type":"string","enum":["low","mid","high"]}}}},"responses":{"NotFound":{"description":"Resource not found","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"Trip not found"}}}}}},"ValidationError":{"description":"Validation failed","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"Validation failed"},"details":{"type":"array","items":{"type":"object","properties":{"path":{"type":"string"},"message":{"type":"string"}}}}}}}}},"InternalServerError":{"description":"Internal server error","content":{"application/json":{"schema":{"type":"object","properties":{"error":{"type":"string","example":"Failed to process request"}}}}}}}}}